BUILD:MACOSX: Simplify building the DockTile plugin
The code now builds the plugin only for the same architecture than
the one for which scummvm is built. This may fail in some cases, as
the plugin need to be the same arch as the system Dock:
 - running ScummVM Intel 64 bits on very early 64 bits Mac OS X systems
   that still had some components, including the Dock,  in 32 bits.
 - running ScummVM Intel 32 bits in more recent full 64 bits system.
The previous build code tried to handle this, but broke in other cases,
such as building for pre-10.14 macOS using a 10.14 SDK, as it tried,
and failed, to build the 32 bit plugin,
The component is not critical, so I prefer to simplify the build
process and ensure it does not fail in some configurations.